# Conflicts:
#	src/main/kotlin/dsx/bps/connection/rpc/JsonRpcHttpClient.kt
#	src/main/kotlin/dsx/bps/core/datamodel/Currency.kt
#	src/main/kotlin/dsx/bps/crypto/CoinsManager.kt
#	src/main/kotlin/dsx/bps/crypto/common/Coin.kt
#	src/main/resources/TestBpsConfig.yaml

+++ src/main/kotlin/dsx/bps/core/datamodel/TxId.kt
@@ -2,5 +2,5 @@ package dsx.bps.core.datamodel
 
 data class TxId(
     val hash: String,
-    val index: Int
+    val index: Long
 )
